OrlandoDiaz_Ex8
I played around with the idea of the particles and made two sources of
“rain” (aka the particles) flood the screen with a little text thrown
in. It took me a while but some classmates helped and the p5.js
examples helped tremendously. Anyway, I was gonna mention something
else but I forgot and it’s 2:00AM. So maybe I’ll remember tomorrow and
post what I was gonna say on slack. Anyway, hope you like it!