Exercise 8
Originally tried to have the particles be an image. I don’t think
there’s anything wrong with my code, since it throws me an error from
the libraries folder. I commented out the areas in which I load the
image and call upon it, for the sake of running the program. I, instead
made simple shapes to show what would have been the images.